refactor: moved transaction generation to nw mod
