+++ jarust/src/jaconnection.rs
@@ -6,9 +6,9 @@ use crate::jasession::JaSession;
 use crate::jasession::WeakJaSession;
 use crate::nw::nwconn::NetworkConnection;
 use crate::nw::nwconn::NwConn;
+use crate::nw::transaction_gen::GenerateTransaction;
+use crate::nw::transaction_gen::TransactionGenerator;
 use crate::prelude::*;
-use crate::transaction_gen::GenerateTransaction;
-use crate::transaction_gen::TransactionGenerator;
 use jarust_rt::JaTask;
 use jarust_transport::trans::TransportProtocol;
 use napmap::UnboundedNapMap;
