+++ jarust/src/nw/mod.rs
@@ -1,4 +1,5 @@
 pub mod nwconn;
+pub mod transaction_gen;
 
 mod demuxer;
 mod jarouter;
